backend/requirements.txt
langchain
langchain-core
langchain-community

google-generativeai
langchain-google-genai
google-genai==0.3.0

ddgs
tenacity>=9.0.0
requests==2.32.4

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1


#for rag based tool
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
numpy==1.26.4
pymupdf==1.24.9
python-dotenv==1.0.1
jinja2==3.1.4
rank-bm25==0.2.2
google-genai==0.3.0
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-google-genai==1.0.8
pinecone==8.1.0
langchain-pinecone==0.1.3
flashrank==0.2.8